Refactor

+++ src/utility.rs
@@ -1,10 +1,16 @@
-use crate::CompileToolErrors;
+use crate::{RuntimeVariables, SourcesCopied, errors::CompileToolErrors};
 use std::{
     fs,
     io::ErrorKind,
     path::{Path, PathBuf},
 };
 use walkdir::{DirEntry, WalkDir};
+
+const LINE_SEPARATOR: &str = "\n######################################################\n";
+
+pub fn print_fancy_block(input: &str) {
+    println!("{LINE_SEPARATOR}{input}{LINE_SEPARATOR}");
+}
 
 /// —
 /// # Errors
@@ -61,6 +67,10 @@ pub fn copy_directory(source: &PathBuf, destination: &Path) -> Result<(), Compil
 /// # Errors
 /// _
 pub fn delete_file(file: &PathBuf) {
+    if !file.exists() {
+        return;
+    }
+
     if let Err(e) = fs::remove_file(file) {
         eprintln!(
             "Error while removing previous compilation file {}: {}",
@@ -74,6 +84,10 @@ pub fn delete_file(file: &PathBuf) {
 /// # Errors
 /// _
 pub fn copy_file(from: &PathBuf, to: &PathBuf) {
+    if !from.exists() {
+        return;
+    }
+
     if let Err(e) = fs::copy(from, to) {
         eprintln!(
             "Error while trying to copy {} to {}: {}",
@@ -96,4 +110,45 @@ pub fn copy_file_if_exists(from: &PathBuf, to: &Path) -> Result<(), CompileToolE
     }
 
     Ok(())
+}
+
+/// Which folder to delete
+/// # Panics
+/// _
+pub fn source_folder_conflict_resolver(runtime_variables: &mut RuntimeVariables) {
+    println!(
+        "Enter `d` to delete the internal sources folder - `{}`\n\
+        Or `I` to ignore this message and compile as is.\n\
+        If you input anything else or an empty string - process will be automatically closed.",
+        runtime_variables.paths.compile_dir_sources.display()
+    );
+    let mut input: String = String::new();
+    std::io::stdin()
+        .read_line(&mut input)
+        .expect("error: unable to read user input");
+    if input.ends_with('\n') {
+        input.pop();
+    }
+    if input.ends_with('\r') {
+        input.pop();
+    }
+    // dbg!(&input);
+
+    if input.to_lowercase() == "d" {
+        if let Err(e) = fs::remove_dir_all(&runtime_variables.paths.compile_dir_sources) {
+            eprintln!("Error while trying to resolve source folder conflict: {e}");
+        }
+        return;
+    }
+    if input.to_lowercase() == "i" {
+        runtime_variables.sources_copied_state = SourcesCopied::Ignored;
+        return;
+    }
+
+    // for all other cases - close the process
+    eprintln!(
+        "Terminated with error: Input was empty / incorrect.\n\
+        Decide which source files folder do you want to keep and remove them manually."
+    );
+    std::process::exit(0);
 }
